requirements.txt
flask
flask-jwt-extended
sqlalchemy
python-dateutil
psycopg2
python-dotenv
gunicorn